Fix bug for drawing
Added the variable draw for a boolean context. This gives the
'mouseover' eventListener change the backgroundColor if it is not true.
Revised the sketchbox eventListener to change the draw variable.
Added a body eventListener to change the draw variable to false. This is
to stop drawing when the mouse event 'mouseup' happend outside the
sketchbox.